pyproject.toml
[tool.poetry]
name = "autogoal"
version = "0.8"
authors = ["Suilan Estevez-Velarde <suilanestevez@gmail.com>", "Alejandro Piad-Morffis <apiad@apiad.net>"]
description = "Automatic Generation Optimization And Learning"
license = "MIT"
readme = "Readme.md"
homepage = "https://autogoal.github.io"
documentation = "https://autogoal.github.io"
repository = "https://github.com/autogoal/autogoal"

classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Code Generators",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]

[tool.poetry.dependencies]
python = "3.9.16"
pytest = "^7.2.1"
numpy = "^1.21.6"
psutil = "^5.9.4"
PyYAML = "^6.0"
fastapi = "^0.89.1"
black = "^22.12.0"
enlighten = "^1.11.1"
pandas = "^1.1.3"
pydot = "^1.4.2"
scipy = "^1.5"
termcolor = "^2.2.0"
toml = "^0.10.2"
tqdm = "^4.64.1"
typer = "^0.7.0"
rich = "^13.2.0"
uvicorn = "^0.20.0"
requests = "^2.28.2"
Cython = "^0.29.33"
networkx = "^2.4"
codecov = "^2.1.12"
dill="^0.3.6"

# -------------------- CONTRIBS ------------------- #
#-------- sklearn
yamlable = "^1.1.1"
[tool.poetry.group.sklearn]
optional = true

[tool.poetry.group.sklearn.dependencies]
scikit-learn = "1.0"
sklearn-crfsuite = "^0.3.6"
#-----------------

#-------- nltk
[tool.poetry.group.nltk]
optional = true

[tool.poetry.group.nltk.dependencies]
nltk = "^3.8.1"
gensim = "4.1"

#-----------------
#----------------------------------------------- #


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
